Module advent_of_code_rust::year_2017::day_23::part_two

Actually executing the instructions here is infeasible, so this is a “decompiling” approach to solving the problem. It turns out this is just counting primes in a specific range, stepping by a specific amount.

The decompiling steps to understand that are as follows. The solution is the value of the register h when the program terminates. The instructions are:

| Instruction ID | Instruction | Control flow notes and questions | | 0 | set b 79 | | | 1 | set c b | | | 2 | jnz a 2 | | | 3 | jnz 1 5 | | | 4 | mul b 100 | | | 5 | sub b -100000 | | | 6 | set c b | | | 7 | sub c -17000 | Constant registers: a = 1, c = 124_900. b = 107_900 to start, incremented by 17 at the end of each iteration of Loop 0. | | 8 | set f 1 | Loop 0 starts. | | 9 | set d 2 | | | 10 | set e 2 | Loop 1 starts. | | 11 | set g d | Loop 2 starts. | | 12 | mul g e | | | 13 | sub g b | | | 14 | jnz g 2 | Conditional within Loop 2. | | 15 | set f 0 | When does f get set to 0 for the rest of the iteration of Loop 1 and 2 (it gets reset to 1 at the start of each Loop 0 iteration)? | | 16 | sub e -1 | | | 17 | set g e | | | 18 | sub g b | | | 19 | jnz g -8 | Loop 2 ends. | | 20 | sub d -1 | | | 21 | set g d | | | 22 | sub g b | | | 23 | jnz g -13 | Loop 1 ends. | | 24 | jnz f 2 | Conditional within Loop 0. | | 25 | sub h -1 | This is the answer. How many times is this instruction run (f needs to be 0 in iteration of Loop 0 for this to run)? | | 26 | set g b | To exit (g == 0 at instruction 28), need b == c. | | 27 | sub g c | | | 28 | jnz g 2 | To exit, need g == 0 here. | | 29 | jnz 1 3 | If this instruction is executed, the program terminates. | | 30 | sub b -17 | | | 31 | jnz 1 -23 | Loop 0 ends. |

The answer is the number of times Loop 1 exits with f = 0 by the time it exits with values that meet the instructions 26-29 criteria to terminate the program. If f is ever set to 0 at instruction 15 at any point in Loop 2 for any iteration of Loop 1 for a given iteration of Loop 0, it remains 0.

The program terminates if Loop 0 finishes an iteration with b == c. c is 17_000 higher than b’s starting value (instruction 7), and c does not change and b only changes by being incremented by 17 at the end of each Loop 0 iteration (instruction 30).

In Loop 2, we set g = d * e - b (instructions 11-13), and if g == 0, we set f to 1. So as soon as we find d and e such that d * e == b, h will be incremented that iteration of Loop 0. For each iteration of Loop 0, d and e each start at 2 and are incremented at the end of their respective loops. So this is really looking for any factorization of b other than 1 * b. The answer is the count of the numbers in the sequence of b with the specified step that are not prime.

Structs

Soln