Module advent_of_code_rust::year_2023::day_20::part_two

This requires at least a partial decompilation approach, as a brute force approach sending button pushes has an infeasible runtime. There are 4 loops that operate in parallel. Each loop feeds into a conjunction module which feeds directly and solely into another conjunction module. Those 4 conjunction modules are the sole inputs to a conjunction module, which has only one destination module: the one we are trying to track when it receives its first low pulse.

If the 4 independent loops are periodic (there is a calculable number of button pushes, after which the conjunction module at the end of the loop sends a low pulse, on repeat), the number of button pushes after which all 4 loops will emit a low pulse is the least common multiple of the periods of each loop.

This solution actually calculates the product of the periods, and will thus only work if the periods share no common denominator other than 1. A more generalized improvement on this would calculate the least common multiple instead of the product.

The diagram looks like this:

flowchart TD
    button --> broadcaster
    broadcaster --> %kr
    broadcaster --> %zb
    broadcaster --> %sm
    broadcaster --> %xd
    %kr --> %hh
    %kr --> &vt
    %hh --> %dh
    %hh --> &vt
    %dh --> %kq
    %kq --> %lm
    %lm --> %hn
    %hn --> %qk
    %hn --> &vt
    %qk --> %cb
    %cb --> %hf
    %cb --> &vt
    %hf --> %ch
    %hf --> &vt
    %ch --> %kd
    %ch --> &vt
    %kd --> %nb
    %kd --> &vt
    %nb --> &vt
    &vt --> %dh
    &vt --> %kr
    &vt --> %kq
    &vt --> %lm
    &vt --> %qk
    &vt --> &lz
    &lz --> &bn
    &bn --> rx
    %cf --> %hl
    %cf --> &qt
    %hm --> %jp
    %vr --> &qt
    %vr --> %sl
    %gq --> %hm
    %gq --> &nl
    %sl --> %jx
    %sl --> &qt
    &pl --> &bn
    %kx --> &dq
    %fr --> %qf
    %rh --> %vr
    &dq --> &mz
    &dq --> %ml
    &dq --> %xd
    &dq --> %fb
    &dq --> %xs
    &dq --> %rc
    &dq --> %rt
    %bv --> &nl
    %jv --> %rh
    %jv --> &qt
    %nd --> %hp
    %gj --> %bv
    %gj --> &nl
    %lv --> %xs
    %lv --> &dq
    %sm --> &qt
    %sm --> %nd
    %nt --> %jv
    %jx --> %cf
    %hl --> &qt
    %hl --> %ng
    &qt --> %sm
    &qt --> %rh
    &qt --> %nd
    &qt --> %jx
    &qt --> %nt
    &qt --> &pl
    %bh --> &nl
    %bh --> %fr
    %gx --> %mh
    %gx --> &dq
    %hp --> %nt
    %hp --> &qt
    %rc --> %lv
    &mz --> &bn
    %qf --> %rd
    %qf --> &nl
    %sk --> &nl
    %sk --> %bh
    %rb --> &nl
    %rb --> %sk
    %fb --> %rt
    %mh --> &dq
    %mh --> %kx
    %rt --> %mt
    %xd --> &dq
    %xd --> %fb
    %ml --> %ts
    %mt --> %rc
    %mt --> &dq
    %ts --> %gx
    %ts --> &dq
    %rd --> &nl
    %rd --> %gq
    %zb --> &nl
    %zb --> %rb
    &nl --> %fr
    &nl --> %zb
    &nl --> %hm
    &nl --> &zm
    &zm --> &bn
    %ng --> &qt
    %xs --> %ml
    %jp --> &nl
    %jp --> %gj

Structs

Soln